🎨 style(policy-list.vue): add tooltip to policies without uiMetadata 💄 style(policy-list.vue): add card--size class to q-card.full-width The tooltip is added to policies without uiMetadata to inform the user that the policy was created on Coinstr-CLI. The card--size class is added to the q-card.full-width element to limit the height of the card and add a scrollbar when the content overflows.
🚀 feat(useNostr.js): add subscribeToPolicies function
The subscribeToPolicies function is added to the useNostr composable to allow the user to subscribe to new policies. The function listens to new policy events and updates the policiesArray with the new policy.

🚀 feat(PoliciesScreen.vue): add subscriptionToPolicies function
The subscriptionToPolicies function is added to the PoliciesScreen.vue component to allow the user to subscribe to new policies. The function listens to new policy events